+++ tests/ViewerTest.php
@@ -16,11 +16,15 @@ namespace BakameTest\Period\Visualizer;
 use Bakame\Period\Visualizer\ConsoleConfig;
 use Bakame\Period\Visualizer\ConsoleOutput;
 use Bakame\Period\Visualizer\Label\LetterGenerator;
+use Bakame\Period\Visualizer\Stdout;
 use Bakame\Period\Visualizer\Viewer;
 use League\Period\Datepoint;
 use League\Period\Period;
 use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
+use function fopen;
+use function rewind;
+use function stream_get_contents;
 
 /**
  * @coversDefaultClass \Bakame\Period\Visualizer\Viewer
@@ -32,55 +36,71 @@ final class ViewerTest extends TestCase
      */
     private $view;
 
+    /**
+     * @var resource
+     */
+    private $stream;
+
     public function setUp(): void
     {
-        $this->view = new Viewer();
+        $this->stream = $this->setStream();
+        $this->view = new Viewer(
+            new LetterGenerator('A'),
+            new ConsoleOutput(
+                ConsoleConfig::createFromRandom(),
+                new Stdout($this->stream)
+            )
+        );
+    }
+
+    public function tearDown(): void
+    {
+        fclose($this->stream);
+    }
+
+    /**
+     * @return resource
+     */
+    private function setStream()
+    {
+        /** @var resource $stream */
+        $stream = fopen('php://memory', 'r+');
+
+        return $stream;
+    }
+
+    /**
+     * @param resource $stream
+     */
+    private function getContent($stream): string
+    {
+        rewind($stream);
+
+        /** @var string $data  */
+        $data = stream_get_contents($stream);
+
+        return $data;
     }
 
     /**
      * @covers ::__construct
-     */
-    public function testConstructor(): void
-    {
-        $viewer = new Viewer();
-        self::assertInstanceOf(LetterGenerator::class, $viewer->getLabelGenerator());
-        self::assertEquals(new ConsoleOutput(), $viewer->getOutput());
-    }
-
-    /**
-     * @covers ::getLabelGenerator
-     * @covers ::setLabelGenerator
-     */
-    public function testLabelGenerator(): void
-    {
-        $labelGenerator = $this->view->getLabelGenerator();
-        self::assertInstanceOf(LetterGenerator::class, $labelGenerator);
-        $this->view->setLabelGenerator(new LetterGenerator('bb'));
-        self::assertNotSame($labelGenerator, $this->view->getLabelGenerator());
-    }
-
-    /**
-     * @covers ::getOutput
-     * @covers ::setOutput
-     */
-    public function testOutput(): void
-    {
-        $output = $this->view->getOutput();
-        $this->view->setOutput(new ConsoleOutput(ConsoleConfig::createFromRainbow()));
-        self::assertNotSame($output, $this->view->getOutput());
-    }
-
-    /**
      * @covers ::sequence
+     * @covers ::view
      * @covers ::filterResultLabel
      * @covers ::addLabels
+     * @covers \Bakame\Period\Visualizer\Stdout::colorize
+     * @covers \Bakame\Period\Visualizer\Stdout::writeln
      */
     public function testDisplaySequence(): void
     {
-        $data = $this->view->sequence(new Sequence(
+        $viewer = new Viewer();
+
+        $this->view->sequence(new Sequence(
             new Period('2018-01-01', '2018-01-15'),
             new Period('2018-01-15', '2018-02-01')
         ));
+
+        $data = $this->getContent($this->stream);
 
         self::assertStringContainsString('A    [--------------------------)', $data);
         self::assertStringContainsString('B                               [-------------------------------)', $data);
@@ -88,24 +108,33 @@ final class ViewerTest extends TestCase
 
     /**
      * @covers ::sequence
+     * @covers ::view
      */
     public function testDisplayEmptySequence(): void
     {
-        $data = $this->view->sequence(new Sequence());
+        $this->view->sequence(new Sequence());
+
+        $data = $this->getContent($this->stream);
+
         self::assertEmpty($data);
     }
 
     /**
      * @covers ::intersections
+     * @covers ::view
      * @covers ::filterResultLabel
      * @covers ::addLabels
+     * @covers \Bakame\Period\Visualizer\Stdout::colorize
+     * @covers \Bakame\Period\Visualizer\Stdout::writeln
      */
     public function testDisplayIntersection(): void
     {
-        $data = $this->view->intersections(new Sequence(
+        $this->view->intersections(new Sequence(
             new Period('2018-01-01', '2018-01-15'),
             new Period('2018-01-10', '2018-02-01')
         ));
+
+        $data = $this->getContent($this->stream);
 
         self::assertStringContainsString('A                [--------------------------)', $data);
         self::assertStringContainsString('B                                 [-----------------------------------------)', $data);
@@ -114,15 +143,18 @@ final class ViewerTest extends TestCase
 
     /**
      * @covers ::gaps
+     * @covers ::view
      * @covers ::filterResultLabel
      * @covers ::addLabels
      */
     public function testGaps(): void
     {
-        $data = $this->view->gaps(new Sequence(
+        $this->view->gaps(new Sequence(
             new Period('2018-01-01', '2018-01-10'),
             new Period('2018-01-15', '2018-02-01', Period::EXCLUDE_ALL)
         ), '');
+
+        $data = $this->getContent($this->stream);
 
         self::assertStringContainsString('A         [----------------)', $data);
         self::assertStringContainsString('B                                    (-------------------------------)', $data);
@@ -131,14 +163,17 @@ final class ViewerTest extends TestCase
 
     /**
      * @covers ::sequence
+     * @covers ::view
      * @covers ::addLabels
      */
     public function testSingleUnitIntervalLength(): void
     {
-        $data = $this->view->sequence(new Sequence(
+        $this->view->sequence(new Sequence(
             new Period('2018-01-01', '2018-02-01'),
             new Period('2017-01-01', '2019-01-01', Period::INCLUDE_ALL)
         ));
+
+        $data = $this->getContent($this->stream);
 
         self::assertStringContainsString('A                                  [-)', $data);
         self::assertStringContainsString('B    [----------------------------------------------------------]', $data);
@@ -151,11 +186,13 @@ final class ViewerTest extends TestCase
     public function testDiff(): void
     {
         $config = (new ConsoleConfig())->withColors('white');
-        $view = new Viewer(new LetterGenerator(), new ConsoleOutput($config));
-        $data = $view->diff(
+        $view = new Viewer(new LetterGenerator(), new ConsoleOutput($config, new Stdout($this->stream)));
+        $view->diff(
             new Period('2018-01-01', '2018-02-01'),
             new Period('2017-12-01', '2018-03-01')
         );
+
+        $data = $this->getContent($this->stream);
 
         self::assertStringContainsString('A                           [--------------------)', $data);
         self::assertStringContainsString('B       [----------------------------------------------------------)', $data);
@@ -174,7 +211,10 @@ final class ViewerTest extends TestCase
             Period::around('2016-06-01', '3 MONTHS')
         );
 
-        $data = $this->view->unions($sequence);
+        $this->view->unions($sequence);
+
+        $data = $this->getContent($this->stream);
+
         self::assertStringContainsString('A                                               [--------------------)', $data);
         self::assertStringContainsString('B                                                                 [-) ', $data);
         self::assertStringContainsString('C         [---------)', $data);